@@ -6,24 +6,24 @@
         "build": "vite build"
     },
     "devDependencies": {
-        "@inertiajs/vue3": "^1.0.0",
-        "@tailwindcss/forms": "^0.5.3",
-        "@vitejs/plugin-vue": "^4.0.0",
-        "autoprefixer": "^10.4.16",
-        "axios": "^1.6.2",
-        "laravel-vite-plugin": "^0.8.0",
-        "postcss": "^8.4.18",
-        "tailwindcss": "^3.2.1",
-        "vite": "^4.0.0",
-        "vue": "^3.2.41"
+        "@inertiajs/vue3": "2.0",
+        "@tailwindcss/forms": "^0.5.10",
+        "@vitejs/plugin-vue": "^5.2.1",
+        "autoprefixer": "^10.4.20",
+        "axios": "^1.7.9",
+        "laravel-vite-plugin": "^1.1.1",
+        "postcss": "^8.5.1",
+        "tailwindcss": "^3.4.17",
+        "vite": "^6.0.7",
+        "vue": "^3.5.13"
     },
     "dependencies": {
         "@element-plus/icons-vue": "^2.3.1",
         "@inertiajs/inertia-vue3": "^0.6.0",
-        "debounce": "^2.1.0",
-        "flowbite": "^2.2.0",
+        "debounce": "^2.2.0",
+        "flowbite": "^2.5.2",
         "lodash.debounce": "^4.0.8",
-        "sweetalert2": "^11.10.1",
+        "pinia": "^2.3.0",
         "vuex": "^4.1.0"
     }
 }
