@@ -14,37 +14,37 @@
                 <div class="md:col-span-2">
                     <div class="flex flex-col max-[500px]:items-center gap-3">
                         <p class="font-semibold text-base leading-7 text-black">
-                            {{ item.name }} {{ item.attributes.invoice }}
+                            {{ editTitle(item.dt_typec)}} {{ item.dt_invoice }}
                         </p>
                         <p class="font-normal text-base text-gray-500">
-                            Артикул: {{ item.attributes.cargo }}
+                            Артикул: {{ item.dt_cargo }}
                         </p>
                         <p
                             class="font-normal text-base leading-7 text-gray-500"
                         >
                             Бренд:
-                            {{ item.attributes.fr_code }}
+                            {{ item.fr_code }}
                         </p>
                         <p
-                            class="font-medium text-base leading-7 text-gray-600 transition-all duration-300 group-hover:text-indigo-600"
+                            class="font-medium text-base leading-7 text-gray-600 transition-all duration-300 "
                         >
-                            {{ item.price }} BYN
+                            {{ item.price }}
                         </p>
                     </div>
                 </div>
-                <input-number
-                    @change="changeQuantity"
+                <InputQuantity
                     :quantity="`${item.quantity}`"
+                    :detailId="`${item.dt_id}`"
                 />
                 <div
-                    class="flex flex-col justify-center ml-[50px] items-center"
+                    class="flex md:flex-col mt-6 md:mt-0 flex-row justify-center ml-[50px] items-center"
                 >
                     <p
-                        class="font-bold text-lg mb-2 text-gray-600 transition-all duration-300 group-hover:text-indigo-600"
+                        class="font-bold text-lg mr-2 md:mr-0 md:mb-2 text-gray-600 transition-all duration-300 group-hover:text-green-600"
                     >
-                        {{ item.price * item.quantity }} BYN
+                        {{ parseFloat(item.price) * item.quantity }} BYN
                     </p>
-                    <button @click="deleteFromCart" class="cursor-pointer">
+                    <button @click="store.deleteDetailFromCart(item.dt_id)" class="cursor-pointer">
                         <svg
                             xmlns="http://www.w3.org/2000/svg"
                             width="24"
@@ -70,44 +70,20 @@
     </div>
 </template>
 
-<script>
-import InputNumber from "@/Components/InputNumber.vue";
-import axios from "axios";
-export default {
-    data() {
-        return {
-            quantity: this.item.quantity,
-        };
-    },
-    components: {
-        "input-number": InputNumber,
-    },
-    props: {
+<script setup>
+import { editDetailTitle } from "@/Services/TitleService";
+import { useCartStore} from "@/Store/cartStore.js";
+import InputQuantity from "@/Components/InputQuantity.vue";
+
+const store = useCartStore();
+const props = defineProps({
         item: {},
-    },
-    created() {
-        console.log(this.item);
-    },
-    methods: {
-        deleteFromCart() {
-            axios
-                .delete(`/cart/${this.item.id}`)
-                .then((res) => {
-                    //console.log(res.data);
-                    this.$emit("getItems", res.data);
-                })
-                .catch((err) => console.log(err));
-        },
-        changeQuantity(count) {
-            console.log(`count: ${count}`);
-            axios
-                .put(`/cart/${this.item.id}`, { quantity: count })
-                .then((res) => {
-                    console.log(res);
-                    this.$emit("getItems", res.data);
-                })
-                .catch((err) => console.log(err));
-        },
-    },
-};
+    });
+
+console.log(props.item);
+
+function editTitle(res) {
+    return editDetailTitle(res);
+}
+
 </script>
