@@ -6,7 +6,7 @@
     >
         <div class="w-full">
             <form
-                class="bg-white shadow-md flex flex-col rounded px-8 pt-6 pb-8 mb-4"
+                class="bg-white shadow-md flex flex-col rounded px-8 pt-6 pb-8 "
             >
                 <div class="mb-4">
                     <label
@@ -32,64 +32,64 @@
                     <textarea
                         v-model="currentDescription"
                         placeholder="Ввести описание"
-                        class="shadow appearance-none border border-red-500 rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
+                        class="shadow appearance-none border  rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
                     />
                 </div>
                 <button
-                    @click.prevent="editPost"
+                    @click.prevent = "editPost"
                     class="bg-green-700 rounded-lg text-white px-5 py-2.5 mx-auto text-lg mt-4"
                 >
-                    Изменить
+                    {{ actionTitle }}
                 </button>
             </form>
         </div>
     </modal>
 </template>
 
-<script>
+<script setup>
 import axios from "axios";
-export default {
-    data() {
-        return {
-            currentTitle: this.title,
-            currentDescription: this.description,
-        };
-    },
-    props: {
-        title: {
-            type: String,
-            default: "",
-        },
-        description: {
-            type: String,
-            default: "",
-        },
-        isShow: {
-            type: Boolean,
-            default: false,
-        },
-        postId: {
-            type: Number,
-            default: 0,
-        },
-    },
+import {ref} from "vue";
 
-    methods: {
-        closeModal() {
-            this.$emit("closeModal");
-        },
-        editPost() {
-            axios
-                .put(`/admin/news/${this.postId}`, {
-                    title: this.title,
-                    description: this.description,
-                })
-                .then((res) => console.log(res))
-                .catch((err) => console.log(err));
-        },
+const props = defineProps({
+    title: {
+        type: String,
+        default: "",
     },
-    created() {
-        console.log(this.postId);
+    description: {
+        type: String,
+        default: "",
     },
-};
+    isShow: {
+        type: Boolean,
+        default: false,
+    },
+    postId: {
+        type: Number,
+        default: 0,
+    },
+    actionTitle: {
+        type: String,
+        default: '',
+    }
+})
+
+const emit = defineEmits(['closeModal'])
+
+const currentTitle = ref("")
+const currentDescription = ref("")
+
+const closeModal = () => {
+    emit("closeModal");
+}
+
+const editPost = () => {
+    console.log(props.postId)
+    axios
+        .patch(`/admin/news/${props.postId}`, {
+            title: currentTitle.value,
+            description: currentDescription.value,
+        })
+        .then((res) => console.log(res))
+        .catch((err) => console.log(err));
+}
 </script>
