@@ -3,8 +3,10 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\RegisteredUserRequest;
 use App\Models\User;
 use App\Providers\RouteServiceProvider;
+use App\Services\CartService;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -19,6 +21,9 @@ class RegisteredUserController extends Controller
     /**
      * Display the registration view.
      */
+    public function __construct(protected CartService $cartService)
+    {}
+
     public function create(): Response
     {
         return Inertia::render('Auth/Register');
@@ -29,20 +34,16 @@ class RegisteredUserController extends Controller
      *
      * @throws \Illuminate\Validation\ValidationException
      */
-    public function store(Request $request): RedirectResponse
+    public function store(RegisteredUserRequest $request): RedirectResponse
     {
-        $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|lowercase|email|max:255|unique:'.User::class,
-            'password' => ['required', 'confirmed', Rules\Password::defaults()],
-        ]);
-
         $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
+            'name' => $request->validated('name'),
+            'email' => $request->validated('email'),
+            'phone_number' => $request->validated('phoneNumber'),
+            'password' => Hash::make($request->validated('password')),
         ]);
 
+        $cart = $this->cartService->storeCart($user->id);
         event(new Registered($user));
 
         Auth::login($user);
