@@ -2,39 +2,41 @@
 
 namespace App\Http\Controllers\Cart;
 
+use App\DTO\CartDTO;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CartFormCreateRequest;
+use App\Http\Requests\CartFormUpdateRequest;
 use App\Services\CartService;
+use App\Services\PriceService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use Inertia\Inertia;
+use Inertia\Response;
 
 class CartController extends Controller
 {
     public function __construct(protected CartService $cartService)
-    {
+    {}
 
+    public function index(): Response
+    {
+        return Inertia::render('Cart/Cart', ['items' => $this->cartService->getCartItems(auth()->id())]);
     }
 
-    public function index(Request $request)
+    public function store(CartFormCreateRequest $request)
     {
-        return Inertia::render('Cart/Cart', ['items' => $this->cartService->index($request->cookie())]);
+        return $this->cartService->addToCart(auth()->id(), new CartDTO($request->validated('id'), $request->validated('quantity'), $request->validated('price')));
     }
 
-    public function store(Request $request)
+    public function update(CartFormUpdateRequest $request, int $id): array
     {
-        return $this->cartService->store($request->cookie(), $request);
+        $this->cartService->updateQuantity(auth()->id(), new CartDTO($id, $request->validated('quantity'), 1));
+        return ['items' => $this->cartService->getCartItems(auth()->id())];
     }
 
-    public function update(Request $request, int $id)
+    public function destroy(int $id): array
     {
-        $this->cartService->update($request->cookie(), $id, $request);
-
-        return ['items' => $this->cartService->index($request->cookie())];
-    }
-
-    public function destroy(Request $request, int $id)
-    {
-        $this->cartService->destroy($request->cookie(), $id);
-
-        return ['items' => $this->cartService->index($request->cookie())];
+        $this->cartService->deleteProductFromCart(auth()->id(), $id);
+        return ['items' => $this->cartService->getCartItems(auth()->id())];
     }
 }
