@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
@@ -10,7 +11,6 @@ class Detail extends Model
     use HasFactory;
 
     public $table = 'detail';
-
     protected $fillable = [
         'dt_id',
         'dt_code',
@@ -38,4 +38,9 @@ class Detail extends Model
         'dt_tp_ptype',
         'lt_dt_acode',
     ];
+
+    public function scopeInvoice(Builder $query,string $invoiceCode):void
+    {
+        $query->where('dt_invoice', '=', $invoiceCode);
+    }
 }
