@@ -2,19 +2,17 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Laravel\Scout\Searchable;
+
 
 class Oems extends Model
 {
-    use HasFactory, Searchable;
+    use HasFactory;
 
-    public function toSearchableArray()
+    public function scopeOfCode(Builder $query, string $code):void
     {
-        return [
-            'dt_invoice' => $this->dt_invoice,
-            'dt_oem' => $this->dt_oem,
-        ];
+        $query->where('dt_invoice', '=', $code)->orWhere('dt_oem', '=', $code);
     }
 }
