@@ -1,13 +1,12 @@
 <template>
-    <div class="antialiased bg-gray-50 dark:bg-gray-900">
+    <div class="antialiased bg-gray-50 ">
         <!-- navbar start -->
         <Navbar />
         <!-- end -->
         <!-- Sidebar -->
         <Sidebar />
         <!-- end -->
-        <main class="p-4 md:ml-64 h-auto pt-20">
-
+        <main class="p-4 ml-0  lg:ml-64 h-auto pt-20">
             <slot />
         </main>
     </div>
