@@ -1,13 +1,13 @@
 <template>
-    <tr class="border-b dark:border-gray-700">
+    <tr class="border-b ">
         <th
             scope="row"
-            class="px-4 py-3 font-medium text-gray-900 whitespace-nowrap dark:text-white"
+            class="px-4 py-3 font-medium text-gray-900 whitespace-nowrap "
         >
             {{ post.id }}
         </th>
         <td class="px-4 py-3">{{ post.title }}</td>
-        <td class="px-4 py-3">{{ post.date }}</td>
+        <td class="px-4 py-3">{{ new Date(post.date).toLocaleDateString() }}</td>
         <td class="px-4 py-3">
             {{ post.description }}
         </td>
@@ -17,7 +17,7 @@
             <button
                 :id="`${post.id}-button`"
                 :data-dropdown-toggle="`${post.id}`"
-                class="inline-flex items-center p-0.5 text-sm font-medium text-center text-gray-500 hover:text-gray-800 rounded-lg focus:outline-none dark:text-gray-400 dark:hover:text-gray-100"
+                class="inline-flex items-center p-0.5 text-sm font-medium text-center text-gray-500 hover:text-gray-800 rounded-lg focus:outline-none "
                 type="button"
             >
                 <svg
@@ -34,16 +34,16 @@
             </button>
             <div
                 :id="`${post.id}`"
-                class="hidden z-10 w-44 bg-white rounded divide-y divide-gray-100 shadow dark:bg-gray-700 dark:divide-gray-600"
+                class="hidden z-10 w-44 bg-white rounded divide-y divide-gray-100 shadow "
             >
                 <ul
-                    class="py-1 text-sm text-gray-700 dark:text-gray-200"
+                    class="py-1 text-sm text-gray-700 "
                     :aria-labelledby="`${post.id}-button`"
                 >
                     <li>
                         <button
                             @click="showModal(post.id)"
-                            class="flex w-full py-2 px-4 hover:bg-gray-100 dark:hover:bg-gray-600 dark:hover:text-white"
+                            class="flex w-full py-2 px-4 hover:bg-gray-100 "
                         >
                             Изменить
                         </button>
@@ -51,7 +51,8 @@
                 </ul>
                 <div class="py-1">
                     <button
-                        class="flex w-full py-2 px-4 text-sm text-gray-700 hover:bg-gray-100 dark:hover:bg-gray-600 dark:text-gray-200 dark:hover:text-white"
+                        @click = "deletePost(post.id)"
+                        class="flex w-full py-2 px-4 text-sm text-gray-700 hover:bg-gray-100 "
                     >
                         Удалить
                     </button>
@@ -60,17 +61,17 @@
         </td>
     </tr>
     <news-form
-        @closeModal="isShow = closeModal"
+        @closeModal="isShow = false"
         :show="isShow"
         :title="post.title"
         :description="post.description"
         :postId="post.id"
+        :actionTitle="`Изменить`"
     />
 </template>
 
 <script>
 import NewsEditForm from "@/Shared/Forms/NewsEditForm.vue";
-import { closeModal } from "@/Services/ModalService";
 
 export default {
     data() {
@@ -91,6 +92,11 @@ export default {
         showModal() {
             this.isShow = true;
         },
+        deletePost(id){
+            axios.delete(`/admin/news/${id}`)
+                .then(res => console.log(res.data))
+                .catch(err => console.log(err))
+        }
     },
 };
 </script>
