@@ -2,97 +2,99 @@
     <layout>
         <div class="bg-white py-24 sm:py-32">
             <div class="mx-auto max-w-7xl">
-                <div class="mx-auto max-w-2xl lg:mx-0">
+                <div class="mx-auto w-full">
                     <p class="text-5xl font-bold tracking-tight text-gray-900">
                         О нас
                     </p>
-                    <div class="mt-12 border-t w-[1300px] border-gray-200">
+                <div class="mt-12 w-full mx-auto flex lg:justify-between justify-center  flex-wrap border-t  border-gray-200">
+                    <div>
                         <iframe
                             src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2348.766396578331!2d27.417182076996344!3d53.93589393044868!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x46dbc4bc3e408231%3A0x56bdcf3265bbe0e8!2z0YPQuy4g0JzQuNGA0LAsINCW0LTQsNC90L7QstC40YfQuCwg0JzQuNC90YHQutCw0Y8g0L7QsdC70LDRgdGC0YwgMjIzMDI4!5e0!3m2!1sru!2sby!4v1711972088485!5m2!1sru!2sby"
-                            class="w-[100%] mt-12"
-                            height="450"
+                            class=" w-full h-[450px] border-0"
                             style="border: 0"
-                            allowfullscreen=""
+                            allowfullscreen
                             loading="lazy"
                             referrerpolicy="no-referrer-when-downgrade"
                         ></iframe>
-                        <div>
-                            <div>
-                                <p class="text-lg">
-                                    Склад : Минская область , Минский район ,
-                                    а/г. Ждановичи , ул. Мира 12
-                                </p>
-                                <p
-                                    class="text-5xl font-bold tracking-tight text-gray-900 mt-8"
-                                >
-                                    Контактная информация
-                                </p>
-                                <div class="flex mt-6">
-                                    <div>
-                                        <p class="font-bold text-xl mb-2">
-                                            Телефон:
-                                            <span class="font-thin"
-                                                >+375 29 630-85-35</span
-                                            >
-                                        </p>
-                                        <p class="font-bold text-xl mb-2">
-                                            Email:
-                                            <span class="font-thin"
-                                                >amiauto.minsk@gmail.com</span
-                                            >
-                                        </p>
-                                        <p class="font-bold text-xl mb-2">
-                                            Skype:
-                                            <span class="font-thin"
-                                                >ami.auto</span
-                                            >
-                                        </p>
+                        <p class="text-lg">
+                            Склад : Минская область , Минский район ,
+                            а/г. Ждановичи , ул. Мира 12
+                        </p>
+                    </div>
+                    <div>
+                        <div class="">
 
-                                        <p
-                                            class="flex flex-col font-bold text-xl"
+                            <p
+                                class="text-3xl sm:text-5xl font-bold tracking-tight text-gray-900 mt-8"
+                            >
+                                Контактная информация
+                            </p>
+                            <div class="flex mt-6">
+                                <div>
+                                    <p class="font-bold text-xl mb-2">
+                                        Телефон:
+                                        <span class="font-thin"
+                                            >+375 29 630-85-35</span
                                         >
-                                            Время работы (прием заявок) :
-                                            <span class="font-normal">
-                                                понедельник-пятница - с 9:00 до
-                                                18:00
-                                            </span>
-                                            <span class="font-normal text-xl">
-                                                суббота, воскресенье - выходной
-                                            </span>
-                                        </p>
-                                    </div>
-                                    <!--div>
-                                        <p class="font-bold">
-                                            УНП:
-                                            <span class="font-thin"
-                                                >690662787</span
-                                            >
-                                        </p>
-                                        <p class="font-bold">
-                                            Юридический адрес :
-                                            <span class="font-thin"
-                                                >г.Минск , ул.Пономаренко, д.35А
-                                                , пом. 208</span
-                                            >
-                                        </p>
-                                        <p class="font-bold">
-                                            Р/С:
-                                            <span class="font-thin"
-                                                >BY46TECN30125404500100000000 в
-                                                РКЦ ОАО "Технобанк" г.Минск ,
-                                                Долгиновский тракт 188 ;</span
-                                            >
-                                        </p>
-                                        <p class="font-bold">
-                                            Код банка:
-                                            <span class="font-thin"
-                                                >TECNBY22</span
-                                            >
-                                        </p>
-                                    <div!-->
+                                    </p>
+                                    <p class="font-bold text-xl mb-2">
+                                        Email:
+                                        <span class="font-thin"
+                                            >amiauto.minsk@gmail.com</span
+                                        >
+                                    </p>
+                                    <p class="font-bold text-xl mb-2">
+                                        Skype:
+                                        <span class="font-thin"
+                                            >ami.auto</span
+                                        >
+                                    </p>
+
+                                    <p
+                                        class="flex flex-col font-bold text-xl"
+                                    >
+                                        Время работы (прием заявок) :
+                                        <span class="font-normal">
+                                            понедельник-пятница - с 9:00 до
+                                            18:00
+                                        </span>
+                                        <span class="font-normal text-xl">
+                                            суббота, воскресенье - выходной
+                                        </span>
+                                    </p>
                                 </div>
+                                <!--div>
+                                    <p class="font-bold">
+                                        УНП:
+                                        <span class="font-thin"
+                                            >690662787</span
+                                        >
+                                    </p>
+                                    <p class="font-bold">
+                                        Юридический адрес :
+                                        <span class="font-thin"
+                                            >г.Минск , ул.Пономаренко, д.35А
+                                            , пом. 208</span
+                                        >
+                                    </p>
+                                    <p class="font-bold">
+                                        Р/С:
+                                        <span class="font-thin"
+                                            >BY46TECN30125404500100000000 в
+                                            РКЦ ОАО "Технобанк" г.Минск ,
+                                            Долгиновский тракт 188 ;</span
+                                        >
+                                    </p>
+                                    <p class="font-bold">
+                                        Код банка:
+                                        <span class="font-thin"
+                                            >TECNBY22</span
+                                        >
+                                    </p>
+                                <div!-->
                             </div>
                         </div>
+                    </div>
                     </div>
                 </div>
             </div>
@@ -100,6 +102,5 @@
     </layout>
 </template>
 
-<script>
-export default {};
+<script setup lang="ts">
 </script>
