@@ -1,22 +1,27 @@
 <template>
-    <section class="bg-white dark:bg-gray-900">
+    <section class="bg-white flex justify-center">
         <div
-            class="grid max-w-screen-xl py-8 mx-auto lg:gap-8 xl:gap-6 lg:py-16 lg:grid-cols-12"
+            class="flex flex-col py-8  sm:justify-center"
         >
-            <div class="mr-auto place-self-center lg:col-span-7">
-                <h1
-                    class="max-w-3xl mb-12 text-4xl font-extrabold tracking-tight leading-none md:text-5xl xl:text-6xl dark:text-white"
-                >
-                    Большой выбор стартеров, генераторов и других автозапчастей
-                </h1>
-                <p
-                    class="max-w-2xl mb-6 font-normal text-gray-500 lg:mb-8 md:text-lg lg:text-2xl dark:text-gray-400"
-                >
-                    Узнайте больше о наших запчастях
-                </p>
+            <div class="flex flex-wrap justify-center xl:justify-between">
+                <div class="flex flex-col">
+                    <h1
+                        class="max-w-4xl xl:max-w-3xl mb-6 sm:text-5xl text-4xl font-extrabold tracking-tight leading-none "
+                    >
+                        Большой выбор стартеров, генераторов и других автозапчастей
+                    </h1>
+                    <p
+                        class="max-w-2xl mb-6 font-normal text-gray-500 lg:mb-8  text-2xl dark:text-gray-400"
+                    >
+                        Узнайте больше о наших запчастях
+                    </p>
+                </div>
+                <img src="../../../../public/starters.jpg" class="sm:w-[350px] sm:justify-between mb-6" alt="mockup" />
+            </div>
+            <div class="flex flex-wrap items-center justify-center xl:justify-start">
                 <a
                     href="/catalog/starters"
-                    class="inline-flex items-center justify-center px-16 py-6 mr-3 text-xl font-medium text-center text-white rounded-lg bg-green-700 hover:bg-green-900 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900"
+                    class="mt-2 sm:mt-0 inline-flex items-center justify-center px-16 py-6 mr-3 text-xl font-medium text-center text-white rounded-lg bg-green-700 hover:bg-green-900 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900"
                 >
                     Стартеры
                     <svg
@@ -34,7 +39,7 @@
                 </a>
                 <a
                     href="/catalog/generators"
-                    class="inline-flex cursor-pointer items-center justify-center px-16 py-6 text-xl font-medium text-center text-gray-900 border border-gray-300 rounded-lg hover:bg-gray-100 focus:ring-4 focus:ring-gray-100 dark:text-white dark:border-gray-700 dark:hover:bg-gray-700 dark:focus:ring-gray-800"
+                    class="mt-2 sm:mt-0 inline-flex cursor-pointer items-center justify-center px-16 py-6 text-xl font-medium text-center text-gray-900 border border-gray-300 rounded-lg hover:bg-gray-100 focus:ring-4 focus:ring-gray-100 dark:border-gray-700  dark:focus:ring-gray-800"
                 >
                     Генераторы
                     <svg
@@ -51,9 +56,7 @@
                     </svg>
                 </a>
             </div>
-            <div class="hidden ml-12 lg:col-span-5 lg:flex">
-                <img src="../../../../public/starters.jpg" alt="mockup" />
-            </div>
         </div>
     </section>
 </template>
+
