@@ -7,12 +7,6 @@
         >
             <div class="mr-4">
                 <img
-                    v-if="detail.dt_foto"
-                    :alt="detail"
-                    class="w-10 h-10 object-cover object-center"
-                />
-                <img
-                    v-else=""
                     src="../../../public/build/no-photo--lg.png"
                     alt="#"
                     class="w-10 h-10 object-cover object-center"
@@ -21,7 +15,6 @@
             <div class="flex flex-col">
                 <p class="font-manrope font-semibold text-lg text-black mb-2">
                     <a
-                        use:inertia-vue3
                         :href="`../../catalog/product/${detail.dt_invoice}`"
                     >
                         <span aria-hidden="true" />
@@ -43,13 +36,9 @@
     </div>
 </template>
 
-<script>
-export default {
-    props: {
-        details: Array,
-    },
-    created() {
-        console.log(!this.details);
-    }
-};
+<script setup>
+
+const props = defineProps({
+    details: Array,
+})
 </script>
