@@ -1,7 +1,7 @@
 <template>
     <form>
-        <div class="flex flex-wrap">
-            <div class="relative w-[60%]">
+        <div class="flex flex-wrap justify-around">
+            <div class="relative w-full md:w-[60%]">
                 <input
                     type="search"
                     id="search-dropdown"
@@ -10,16 +10,39 @@
                     v-model="searchQuery"
                     @input="getSearchingDetails"
                 />
+                <div v-if="categoryList.length !== 0"  class="absolute top-[45px] rounded-[15px]  z-10 w-full h-[200px] overflow-y-auto">
+                    <div  class="bg-white  pt-4  border-b-6 border-gray-300 ">
+                        <div class="overflow-hidden">
+                            <p class="font-bold text-xl px-4">Категории</p>
+                        </div>
+                        <div v-for="category in categoryList" class="overflow-hidden border-t-2 p-4 border-gray-300">
+                            <inertia-link
+                                :href="`${otherParts.get(`${category}`)}`"
+                                class=" text-2xl p-8 border-b-gray-300"
+                            >
+                                {{ category }}
+                            </inertia-link>
+                        </div>
+                    </div>
+                </div>
                 <div
                     class="absolute top-[45px] rounded-[15px]  z-10 w-full h-[200px] overflow-y-auto"
                     v-if="details.length !== 0"
                 >
                     <!--detail-list :details="details" /!-->
-                    <div v-for="detail in details" class="bg-white grid grid-cols-3 gap-4 p-2  border-b-2 border-gray-300">
-                        <a :href="`/catalog/product/${detail.dt_code}`" class="text-xl p-4">{{detail.dt_code}}</a>
-                        <p class="text-xl p-4">{{detail.dt_firm}}</p>
-                        <p class="text-xl p-4">{{detail.dt_typec}}</p>
+                    <div  class="bg-white  pt-4  border-b-2 border-gray-300">
+                        <div class="grid grid-cols-3 gap-4 overflow-hidden">
+                            <p class="font-bold text-lg px-4">Код</p>
+                            <p class="font-bold text-lg ">Бренд</p>
+                            <p class="font-bold text-lg px-4">Наименование</p>
+                        </div>
+                        <div v-for="detail in details" class="grid grid-cols-3 gap-4 overflow-hidden border-t-2 border-gray-300">
+                            <a :href="`/catalog/product/${detail.dt_code}`" class="text-xl p-4 border-r-2 border-gray-300">{{detail.dt_code}}</a>
+                            <p class="text-xl p-4 border-r-2 border-gray-300">{{editTitle(detail.dt_firm)}}</p>
+                            <p class="text-xl p-4 border-r-2 border-gray-300">{{editTitle(detail.dt_typec)}}</p>
+                        </div>
                     </div>
+
                 </div>
                 <inertia-link
                     :href="`/catalog/search?searchQ=${searchQuery}`"
@@ -43,7 +66,7 @@
                     <span class="pl-2 text-md">Найти</span>
                 </inertia-link>
             </div>
-            <div class="flex space-x-2 ml-[80px]">
+            <div class="flex space-x-2 mt-2">
                 <svg
                     class="w-6 h-6 text-white dark:text-white"
                     aria-hidden="true"
@@ -64,39 +87,49 @@
     </form>
 </template>
 
-<script>
-import axios from "axios";
-import debounce from "lodash.debounce";
+<script setup>
+import { ref, computed } from 'vue';
+import axios from 'axios';
+import debounce from 'lodash.debounce';
+import { editDetailTitle } from '@/Services/TitleService';
+import { otherParts } from '@/Store/index.js';
 
-export default {
-    props: {
-        link: {
-            type: String,
-            default: "",
-        },
+const props = defineProps({
+    link: {
+        type: String,
+        default: "",
     },
-    data() {
-        return {
-            searchQuery: "",
-            details: [],
-            search: '',
-        };
-    },
-    methods: {
-        getSearchingDetails: debounce(function () {
-            if (this.searchQuery === "") {
-                this.details = [];
-            } else {
-                axios
-                    .get(`${this.link}=${this.searchQuery}`)
-                    .then((res) => {
-                        console.log(res.data.details);
-                        this.details = res.data.details;
-                        this.search = res.data.search;
-                    })
-                    .catch((err) => console.log(err));
-            }
-        }, 1000),
-    },
-};
+});
+
+const searchQuery = ref('');
+const details = ref([]);
+const search = ref('');
+const categoryList = ref([]);
+
+const otherPartsData = computed(() => otherParts);
+
+const getSearchingDetails = debounce(() => {
+    let categories = Array.from(otherPartsData.value.keys()).filter(key => key.includes(searchQuery.value));
+    categoryList.value = Array.from(categories);
+    if (searchQuery.value === "") {
+        details.value = [];
+        categoryList.value = [];
+    } else {
+        axios
+            .get(`${props.link}=${searchQuery.value}`)
+            .then((res) => {
+                console.log(res.data.details);
+                details.value = res.data.details;
+                search.value = res.data.search;
+            })
+            .catch((err) => console.log(err));
+    }
+}, 1000);
+
+const editTitle = (res) => editDetailTitle(res);
 </script>
+
+
+
+
+
