@@ -1,13 +1,13 @@
-<template>
+<template class="w-full">
     <Header> </Header>
-    <main scroll-region>
+    <main class="w-full">
         <slot />
     </main>
     <Footer></Footer>
 </template>
 
 <script setup>
-import Header from "./Header.vue";
+import Header from "./Header/Header.vue";
 import Footer from "./Footer.vue";
 defineProps({
     details: Object,
