@@ -1,11 +1,16 @@
 <?php
 
+use App\Http\Controllers\Admin\AdminApproveUserController;
 use App\Http\Controllers\Admin\AdminAuthController;
+use App\Http\Controllers\Admin\AdminCurrencyController;
+use App\Http\Controllers\Admin\AdminUserController;
+use App\Http\Controllers\Cart\ClearCartController;
+use App\Http\Controllers\Desktop\DesktopController;
 use App\Http\Controllers\Admin\AdminController;
 use App\Http\Controllers\Admin\AdminSearchController;
-use App\Http\Controllers\Admin\DetailController;
-use App\Http\Controllers\Admin\NewsAdminController;
-use App\Http\Controllers\Admin\OrderAdminController;
+use App\Http\Controllers\Admin\AdminDetailController;
+use App\Http\Controllers\Admin\AdminNewsController;
+use App\Http\Controllers\Admin\AdminOrderController;
 use App\Http\Controllers\Cart\CartController;
 use App\Http\Controllers\Catalog\BearingController;
 use App\Http\Controllers\Catalog\CatalogSearchedController;
@@ -60,8 +65,10 @@ Route::group(['prefix' => 'catalog'], (function () {
 }));
 
 Route::get('/info', [InfoController::class, 'index'])->name('info.index');
+Route::get('/desktop', [DesktopController::class, 'index'])->name('desktop.index');
 
 Route::middleware('auth')->resource('/cart', CartController::class)->only(['index', 'store', 'update', 'destroy']);
+Route::middleware('auth')->put('/clear', [ClearCartController::class, 'index'])->name('cart.clear');
 Route::middleware('auth')->resource('/order', OrderController::class)->only(['index', 'store', 'update', 'show']);
 
 Route::group(['prefix' => 'admin', 'middleware' => 'RedirectIfAdmin'], function () {
@@ -74,20 +81,33 @@ Route::middleware(['auth', 'admin'])->prefix('admin')->group(function () {
     Route::get('/dashboard', [AdminController::class, 'index'])->name('admin.dashboard');
 
     // details route
-    Route::get('/details', [DetailController::class, 'index'])->name('admin.details.index');
-    Route::post('/details/store', [DetailController::class, 'store'])->name('admin.details.store');
-    Route::put('/details/update/{id}', [DetailController::class, 'update'])->name('admin.details.update');
-    Route::delete('/details/delete/{dt_id}', [DetailController::class, 'delete'])->name('admin.details.delete');
+    Route::get('/details', [AdminDetailController::class, 'index'])->name('admin.details.index');
+    Route::post('/details/store', [AdminDetailController::class, 'store'])->name('admin.details.store');
+    Route::put('/details/update/{id}', [AdminDetailController::class, 'update'])->name('admin.details.update');
+    Route::delete('/details/delete/{dt_id}', [AdminDetailController::class, 'delete'])->name('admin.details.delete');
 
     // orders route
-    Route::get('/orders', [OrderAdminController::class, 'index'])->name('admin.orders.index');
+    Route::get('/orders', [AdminOrderController::class, 'index'])->name('admin.orders.index');
+    Route::get('/orders/{order}', [AdminOrderController::class, 'show'])->name('admin.orders.show');
 
     // news route
-    Route::get('/news', [NewsAdminController::class, 'index'])->name('admin.news.index');
-    Route::put('/news/{post}', [NewsAdminController::class, 'update'])->name('admin.news.update');
-
+    Route::get('/news', [AdminNewsController::class, 'index'])->name('admin.news.index');
+    Route::post('/news/store', [AdminNewsController::class, 'store'])->name('admin.news.store');
+    Route::patch('/news/{post}', [AdminNewsController::class, 'update'])->name('admin.news.update');
+    Route::delete('/news/{post}', [AdminNewsController::class, 'destroy'])->name('admin.news.delete');
     Route::get('/api/search', [AdminSearchController::class, 'index'])->name('admin.search.index');
 
+    //users route
+    Route::get('/users', [AdminUserController::class, 'index'])->name('admin.users.index');
+    Route::get('/users/{user}', [AdminUserController::class, 'show'])->name('admin.users.show');
+    Route::put('users/{user}', [AdminUserController::class, 'update'])->name('admin.users.update');
+    Route::delete('/users',[AdminUserController::class, 'destroy'])->name('admin.users.delete');
+
+    //currency route
+    Route::get('currency', [AdminCurrencyController::class, 'index'])->name('admin.currency.index');
+    Route::post('currency', [AdminCurrencyController::class, 'update'])->name('admin.currency.update');
+
+    Route::put('/approve/{user}', [AdminApproveUserController::class, 'index'])->name('admin.approve.index');
 });
 
 require __DIR__.'/auth.php';
